Revert "add 10-request-form"
