+++ package.json
@@ -13,7 +13,6 @@
     "@popperjs/core": "^2.11.6",
     "@types/bootstrap": "^5.2.6",
     "@types/lodash": "^4.14.191",
-    "@types/nodemailer": "^6.4.8",
     "@typescript-eslint/eslint-plugin": "^5.57.1",
     "@typescript-eslint/parser": "^5.57.1",
     "bootstrap": "^5.3.0",
@@ -32,7 +31,6 @@
   },
   "dependencies": {
     "lodash": "^4.17.21",
-    "nodemailer": "^6.9.3",
     "vue3-carousel": "^0.2.16"
   }
 }
