+++ pages/index.vue
@@ -43,7 +43,6 @@
     <our-advantages class="mb-5" />
     <our-workflow class="mb-5" />
     <partners class="mb-5" />
-    <request-form class="mb-5" />
   </div>
 </template>
 
@@ -55,8 +54,6 @@ import ProductsList from '~/components/productsList/ProductsList.vue'
 import OurTechnics from '~/components/ourTechnics/OurTechnics.vue'
 import AboutCompany from '~/components/aboutCompany/aboutCompany.vue'
 import Partners from '~/components/partners/Partners.vue'
-
-const { data } = await useFetch('/api/hello')
 </script>
 
 <style lang="scss" scoped>
