--- server/api/hello.ts
@@ -1,5 +1,0 @@
-export default eventHandler((event) => {
-  return {
-    hello: 'world'
-  }
-})
