--- server/index.ts
@@ -1,4 +1,0 @@
-import { Nitro } from 'nitropack'
-export default async (_nitro: Nitro) => {
-
-}
